+++ docs/src/news.md
@@ -4,7 +4,7 @@
 
 - Breaking
     - **miss**: renamed to **findmiss**.
-    - **missdf** Declare data as missing in a data frame.
+    - **recod_miss** Declare data as missing in a data frame.
 
 - Modifications
     - Code cleaning.